+++ service/src/test/kotlin/no/nav/su/se/bakover/service/utbetaling/StansUtbetalingServiceTest.kt
@@ -303,7 +303,7 @@ internal class StansUtbetalingServiceTest {
 
     private val sak: Sak = Sak(
         id = sakId,
-        saksnummer = Saksnummer(Math.random().toInt()),
+        saksnummer = Saksnummer(Math.random().toLong()),
         opprettet = Tidspunkt.now(),
         fnr = fnr,
         oppdrag = mock()
